refactor(extraction): move grounding orchestration
